Added ArrayUnionMatcher as optimization for OR queries. Changed Query.matcher() signature.
The new OR matcher is several times faster (the number of times faster increases with the
number of sub-clauses), but does not allow access to the "current document". The new
context argument allows the Or query to know when it can use the optimized matcher.

I needed a way during instantation of the matcher tree to let the query know whether access
to the current document was needed. Rather than add another keyword argument to
Query.matcher(), I replaced the weighting argument with a "context" argument that subsumes
weighting and the new "needs_current" functionality, which also replaces the
"requires_matcher" signal from categorizers to collectors.

See issue #271.